enotif: Handle Talk notifications via new Notifications system
Migrated second ENotif Talk mail into Notification system. This
is a third iteration. At the end all different notifications cause
exactly the same Email.
To future proof the system, we can pass the Source, which means
what triggered this notification. Is it Talk, is it Watchlist or
is caused by fact that Recipient is on the ENotifAll list.

Bug: T387995
Change-Id: I1993208882097c591f93fb47d80076a93c34fcd9