ResourceLoader: Test for WikiModule titleinfo cache and purge
Some of this is covered by PageUpdaterTest (via ResourceLoaderUpdateSpyTrait)
but this hardcoded internal implementation datails about WANObjectCache
and even specific cache keys, and does not actually confirm that
this key is correct, or otherwise that it results in the needed
behaviour.

As a first step toward refactoring and simplifying this, first
write a high-confidence observational test that validates that caches
are used instead of fetching new data, and that after editing new
data comes through - without any mocking or inspecting of internal
calls.

Bug: T393835
Change-Id: Ice1ad07303f12e5602b33d4de7500791448b8f18