logging: Always declare variables in RightsLogFormatter group formatting
Setting conditional variable (and use it under the same conditional),
makes it hard for phan to see the correct type for the variables.
phan fails with "Suspicious array access" when formatDate() gets a
return type declaration `array`.

Unpack the array in the foreach and use the same variable to check.

Change-Id: I8408d10774cdf96ee59c0f1de3729aa41d2602b6