Stats: Remove not needed key normalization
The first parameter is almost always a string that's hard-coded in
the code. Here are some example usages:
https://codesearch.wmcloud.org/search/?q=%3EsetLabel%5C%28&files=%5C.php%24&repos=MediaWiki+core

There is not much value in normalizing these. It's actually more
confusing if a key from the code cannot be found in the database
because it has been tinkered with. Instead let the validateLabelKey
function throw an exception. This allows us to fix errors in the
code – if there even are any.

This was suggested in I25132ac.

Bug: T378764
Change-Id: I9ce0cbf5bbabf8b77da6b32fa99f1ed7ac6c7094