SpecialBlock [Codex]: Include an 'Active range blocks' table for IPs
When the target is an IP or IP range, show an 'Active range blocks'
accordion that lists all range blocks that apply to the target. The idea
is that admins should be informed if an IP or range is already blocked,
even if the target is a wider CIDR range. Each row in the table has a
'View' link to load that range as the target. We do this instead of
providing direct 'Edit' / 'Remove' links because admins may not realize
the range block is for a different target.

In effect, 'Active range blocks' surfaces blocks that effect the target,
but not directly. The 'Active range blocks' accordion is always shown
for IPs and ranges (situations where there could be active range blocks)
but has an empty state like the other tables.

Bug: T389987
Change-Id: I4d5f42984b05cf1b30e2f962b9381779196fcab3