media: Use `getWithSetCallback()` on WANObjectCache instead
The `get()` and `set()` methods in WANObjectCache are meant to be
internal and not recommended to be used directly. So for cases where
we want to read some cached data and recompute incase it's not found,
it's recommeded to use `getWithSetCallback()`.

This patch migrates usage to the recommended form. Ahead of this effort
is to mark the aforementioned methods as internal.

Change-Id: I81fc8a8e6f40c3fe09ea1308684c10ce67ec680d