Stats: Create UnitTestingHelper::consumeAllFormatted() to ease assertions
Developers should generally assert the full response, not just assert
that the response includes a certain piece or part. If we assume that
the code under test is deterministic and well-written, this should
be easy to do, and I think the helper should encourage that.

The existing methods make it hard to do this, since the selectors
don't require matching all labels, so the path of least resistence
is to write a generic selector, and test just the one metric that
you're thinking of while writing the intitial code for a feature.
Then, as more get added, there is no push to cover those in tests,
and thus that new code becomes uncovered by default, and prone to
regressions.

When testing legacy code, or otherwise running higher-level
tests where a lot of code is indirecty executed, or where code suffers
from non-deterministic labels and values, it of course super useful
to be able to assert just the part you need without caring about
other labels, actual values, or other metrics. That way people will
at least write some tests, instead of putting off entirely by requiring
a large refactor first. But, I think that is an exception rather
than the rule.

Ref https://timotijhof.net/posts/2015/qunit-anti-patterns/.

Change-Id: If673404186396feede0373143f3aa9ea08f9bf88