Stats: Remove unreachable $wgStatsPrefix fallback from ServiceWiring
It is conventional in Dogstatsd/Prometheus to use lowercase prefixes
for applications. This is different from the Statsd/Graphite
convention.

The default in MainConfigSchema::StatsPrefix is 'mediawiki', and this
matches WMF production as well.

The inline fallback here of camel-cased "MediaWiki" is unreachable
as it isn't supported to unset or null out configuration variables
at runtime.

Change-Id: I3e60b36b872493e7bc3b0b6d79110778a536f60b