Move Parsoid module configuration array schema to Parsoid repository
Rather than have to maintain *three* parallel schemas for Parsoid
module configuration arrays, remove the property details from the
schema in core and add a pointer to the full documentation.

As of Ib4c94eab0f3091d224c334b92fcb79215c519121, the Parsoid module
configuration schema is now included in the Parsoid repository, and
configurations are validated when Parsoid modules are loaded.

Note that the extension.schema.v1 is considered "frozen" and so it has
not been updated.  The definition for the "annotations" property
underneath a Parsoid module is incorrect in the frozen version of
'v1'; use a 'v2' schema or an ExtensionModule if you need to register
an annotation handler.

Follows-Up: Ib4c94eab0f3091d224c334b92fcb79215c519121
Change-Id: Ib7b61feeb6e368bf0587d91173d519bd639f501e