Rest: Replace ad-hoc code with StatsFactory::newUnitTestingHelper
== Improve debugging ==

While at it, improve the error message for when a selector matches
a metric name, but not any of the labels on its samples. Previously,
there was nothing to figure out what actually happened.

Before:
> OutOfRangeException: Metric selector 'html_input_transform_total{original_html_given=true}' matched zero samples.

After:
> OutOfRangeException: Metric selector 'html_input_transform_total{original_html_given=true}' matched zero samples.
>
> Actual samples:
> mediawiki.html_input_transform_total:1|c|#original_html_given:false,page_exists:true,status:unknown
> mediawiki.html_input_transform_total:1|c|#original_html_given:as_renderid,page_exists:unknown,status:hit_stashed

For the case where the metric name wasn't found at at all,
it previousy printed these as individual lines to the debug log,
but that's quite the sea of stuff and should not be the first place
to find these. I've moved these into the exception message, too.

== Fix incomplete metric match ==

Previously, testSetOriginal() was asserting that the Graphite sub-metric
started with the word "given" which isn't a thing in Prometheus.

"not_given" maps cleanly to "original_html_given:false" but the
case of "given" is either "given=true" or "given=as_renderid". Make
this explicit.

Change-Id: I28007728f2677a5f8963d3b7c62365917cfa2d13