phpunit: Improve assertArrayContains to support lists
The built-in assertContains had terrible debugging diagnostics.
In practice, what you want is for `$this->assert…(['x', 'y'], $actual);`
to do in_array for each item, and when if it returns false of any,
to print the actual array in the place where a diff would normally
appear. Upstream has declined this, and for the case of associative
arrays we have implemented assertArrayContains for that reason, but
it did not yet support lists. This is now fixed.

Change-Id: I7c9ad17847df2265c301c1ed3bd8e40b639a6ca8