Add metadata to OutputPage after running the output pipeline
The output pipeline can modify the metadata of the ParserOutput. For
them to be complete in the OutputPage, they need to be added after the
output transform pipeline is run.

This fixes issue T384151 where the DiscussionTools threads are not
correctly available when returning from editing a thread (where the
rendering is done via ApiParse).

Bug: T384151
Change-Id: I3e677877f5a0c9c4236eba934c7ed6a458495874