Html: use new Codex menu groups feature in listDropdownOptionsCodex()
Before we were faking groups by disabling options. Following T305036, we
should be using Codex's built-in grouping functionality.

This change necessitates updating the Vue application to handle the
groupings of options. While we're in here, clean up the reason parsing
logic and move it to ReasonField.vue which is the only place that should
care about it. SpecialBlock.php still requires special handling to
support the 'wpReason-other' URL parameter.

Fix linting error in UserLookup.vue

Bug: T391417
Change-Id: I084bb2e2973938e4a31a3880fb83a609d283d583