Set spellcheck=false on the searchbox input
This stops webkit based browsers (particularly iOS Safari) from applying
smart quotes, which produce unexpected behavior in search. Someone
using quotes almost certainly wants to be searching for a literal
phrase, not searching for actual curly quotes.

It's also probably sensible to avoid autocorrect in the search input for
other reasons, as people are searching for a lot of names that might
get autocorrected to adjacent dictionary words.

Bug: T385525
Change-Id: I28d0fe5f7df4d3020966b83c89e48712931216f5