Prevent strip markers in HTML when using {{#tag}}
When using {{#tag}} with an unhandled extension tag name, strip markers
in the tag name or attributes could show up in the output HTML, mangled
in such a way that they are not properly replaced.

Additionally, strip markers in attributes to <pre> could show up in
the output; again mangled by encoding such that they are not properly
replaced.

Finally, strip markers in indicator names would show up in the
indicator map.

In all three of these cases, we now use Parser::killMarkers() to
remove strip markers when obtaining string values of the tag name,
<pre> attributes, or <indicator> name.

Bug: T391531
Bug: T387608
Change-Id: Ia8b4efc544a3bc08c17099252593610c9ac41d52