Language::getJsDateFormats(): Always include default format in styles
Always include the default date format in the styles, even if it
is not present in the date preferences array.

On the client side, remove the fallback to "dmy" and fix the incorrect
comment about the default format being the site default.
config.defaultStyle is actually the default in the selected user
language, not the site default. Since both the default and the format
list come from the same Language object, there's no need for
language-independent fallbacks.

Bug: T391423
Change-Id: I89863c41851264fd08add52a5e846b03d5c8cdfb