Make importImages treat same name+SHA as skipped
When --overwrite is specified and --skip-dupes is not specified,
importImages would previously consider an upload as failed if the
current revision of the file matched the same name and the same SHA
hash. This prevents idempotency under those conditions, which can arise
when running the script multiple times against the same set of files.
Declaring them as skipped rather than failed makes sense because the end
result is the expected behavior of the script (the specified file in the
import directory is now the current revision of that file on the wiki),
and prevents us from exiting with a non-zero exit code in such cases.

While --skip-dupes could be used in theory to work around this edge
case, it is not appropriate for the cases where it is desired to have
the same file living at multiple different names.

The error message for failed uploads now also makes use of the error
message produced by the upload's return Status, if one exists.

Change-Id: I778a7a477e74f38f895d80e581087b68ecda4607