Assert that hooks return void iff not abortable
The general expectation is that a hook method returns void if and only
if the hook runner calls the hook with 'abortable' => false. Add
assertions in both directions; but because not all hooks match this
expectation at the moment (partly due to oversight because this wasn’t
enforced before, I expect), also allow test subclasses to opt individual
hooks out of these assertions.

Change-Id: I858d5dc02c65a0c68081214c4c71a7d816c0a524
Depends-On: I7395ce040212ac67c1a895b7f53d3f1926b7b716
Depends-On: I99971a1b1616cb3f8ef6ea3057e6cc1ecdcfc5a2
Depends-On: Ibde07eeef76bec9c14f34d8ce99f74bf22108b0d
Depends-On: If37b72896f18d751bfaa19e6d340881319587b5f