UndeletePage: fix before-state in PageRevisionUpdatedEvent
Why:
- When undeleting revisions into an existing page in a way that replaces
  the latest revision, the PageRevisionUpdatedEvent must correctly
  represent the state of the page prior to undeletion.

What:
- Call DerivedPageDataUpdater::grabCurrentRevision() in UndeletePage before
  manipulating the page table.
- Fixed improper timestamp comparison in UndeletePage::undeleteRevisions()
  so the new tests can pass.

Note: This patch does not directly address T391739 as reported. The
cause of the error is still unknown. It just fixes an inconsistency in
related code, in the hope that it may avoid similar issues in the
future. If we are lucky, it may end up fixing the error at hand as
well.

Bug: T391739
Change-Id: Ia6f9731c6b2bfd7d19a46564b4b6248230d623d5