specials: Use mime type from database on Special:UploadStash
Allow the UploadStashFile to get the mime type and pass to the parent
constructor.
The mime type is used on Special:UploadStash to render a thumb for the
file in the list, if not provided it is calculate from the raw file,
transfered from swift. For many big files in the list this can timeout.

Also provide the sha1 as needed for the next argument,
but it is unused at the moment on the special page.

Also set the new argument in UploadStash::initFile

Bug: T381109
Change-Id: I664c028f84d4b50ceb5bc0999a9dea02fb1395b1