Add a QueryTrackingCategories list module to the action API
This patch adds a trackingcategories list module to the action API, to
be able to get information and generate list of tracking categories as
defined in [[Special:TrackingCategories]], if they exist (i.e. if they
contain pages or if their category page exists).

It refactors ApiQueryAllCategories since a significant part of the code
can be re-used between these two modules.

Bug: T76290
Change-Id: I640104e13d3ab00ebd22b9c195f844827f4638df