diff.js: add documentation and fix unlocking
Add doc comments to a few functions, most notably
`maybeClearSelectProtection` whose behaviour is not at all obvious.

Pass null to jQuery's .attr to unset the attribute. I'm not sure if
undefined has ever worked, but it definitely doesn't today. I have a
vague memory of coming across this in the past, but I don't remember if
it was for this code, and what was the conclusion.

By fixing the above, a bug emerged in Firefox where the locking would be
lost during a selection as soon as the pointer exits a diff cell. This
is caused by Firefox updating the Selection's `anchorNode` to the outer
`<tr>` element, which does not belong to any side. In turn, this caused
the side-locking to be cleared. This was already the case before this
patch, but did not cause breakage precisely because the lock-clearing
logic was non-functional to begin with.

So, switch to listening to `selectstart` events rather than
`selectionchange`. The locked side is not meant to change
mid-selection, so we only need to act once per selection, which the
`selectstart` event lets us do, while also improving performance by
avoiding pointless runs of the handler. This change, in turn, forces us
to delay running the side-locking code, using the `setTimeout( fn, 0 )`
trick: without it, Chrome would not have access to the current Selection
object when calling `getSelection()`.

Also, do not clear selection when no side is locked. It's pointless.
Even clearing it when a side is locked is questionable, and it seems to
produce bugs (T391168). While this patch does not fix those bugs, it
should make it easier to reason about them and validate the fixes.

Tested in Chrome and Firefox, quite thoroughly, but it might be
desirable to have proper QA once all the patches for T391168 have been
merged.

Bug: T391168
Change-Id: Ie90c8957bef907acecb8a6bb6274a047f2fe52ef