ParsoidParser: use consistent ParserOptions in ::parse and PageConfig
Instead of manually transferring options between the
ParserOptions provided to ::parse() and the one held by
PageConfig, just reuse the same object.

Reusing the same ParserOptions for subdocument processing to
different ParserOutputs does require DataAccess to save/restore
the watcher which links the now-canonical ParserOptions to the
"current" ParserOutput.

Deprecate PageConfigFactory::create() in the process, in favor of
a new PageConfigFactory::createFromParserOptions() method that
allows the caller to pass the desired ParserOptions object.

We also clean up an outdated comment about the limit report
setting, and as mentioned above take greater care in DataAccess
not to clobber the watchers set on the ParserOptions.  This patch
adds `currentRevisionRecordCallback` to the set of common Parsoid
"used options", since it is used in `Parser::makeLimitReport()`
to record the size of the input wikitext.
`ParserOptions::isSafeToCache()` ignores
`currentRevisionRecordCallback` so adding this as a used option
is harmless.

This is an alternative to
I7ac9182bb3121f277f2311c3aa321c341119e295.

Bug: T392113
Change-Id: I5e2c789d12d444f582238fdde09236f397896ef6