Maintenance: Ensure StatsFactory flushes without legacy StatsD
As of Icb8f08f7149c6, StatsFactory::flush is safe to call frequently
because it no longer does any work if the cache is empty.

This is called via MediaWikiEntryPoint::emitBufferedStats.

Note that Maintenance::output() is already a debounced yield signal
in its own right, because scripts generally either do meaningful
chunks of work between prints, or, if printing a multi-line summary,
do nothing that would emit stats. As such "zero" is a fine treshold
to do a flush if/when flush() is called. No additional batching is
needed.

This matches what we have already done for a long time with other
yield signals such as Maintenance::commitTransactionRound,
which already didn't have any threshold or condition on top of
the fact that a database transaction is being committed.

== Legacy statsd ==

Per the above, StatsFactory::flush now includes a check within it,
but legacy BufferingStatsdDataFactory has no such thing. This is okay
because we don't flush that directly today. Instead, we call
MediaWikiEntryPoint::emitBufferedStats, which has its own hasData()
check already before it instructs BufferingStatsdDataFactory to
do any work.

Bug: T381042
Depends-On: Ib490666ffa4bf87234abf156bd5ae1a5ad9282db
Change-Id: I10dcd8db98531d63de1d9188453bba5e9289af54