Improve docs for $wgLocalDatabases
The phrasing "site" and "developer account" are outdated in this context,
and may create confusion to the modern reader.

== History ==

This phrasing originates from r2579 in 2004 (2c34ad2fc0), two full
years before CentralAuth would be created in 2006.

$wgLocalDatabases was first used to improve an early cross-wiki
user rights feature, in the form of Special:Makesysop. This was
limited to developers/sysadmins and took input in the form `Example@en`,
mapped directly to a database name in the form "en" . "wiki".
$wgLocalDatabases restricted and validated the inputs.

This was replaced with Special:Userrights by r28650 in 2007 (c3ef834c04),
by which time it took wiki IDs of the form "enwiki" directly from the
user interface, thus making space for wiki families like Wikibooks,
and Wiktionary which have their own "en" wiki.

Change-Id: I62fb50a6919cedd252166c386386d51c0a20f472