UserLinkRenderer: Convert mw-tempuserlink-expired tooltip to plain CSS (take 2)
== Ideal ==

* Native `title` attribute.
* Positioned automatically correctly near edge of screen,
  as native tooltip switch from left-aligned to right-aligned if the
  element is near the edge of the screen.
* Positioned automatically correctly near edge of window,
  as native tooltips are outside and on top of the browser window,
  they can thus extend beyond the edge of the window.

  https://phabricator.wikimedia.org/F59291036
* Accessible.
* Hidden on Esc, in a way that retains focus/tabindex position.
* Styled with Codex.

This is not yet possible, but in the future browsers will hopefully
expose the native `::tooltip` pseudo element, to allow it to be
styled. https://github.com/w3c/csswg-drafts/issues/8930

== Next best thing ==

* Generate the tooltip with pure CSS and no extra JS or HTML.
* Hide and show based on CSS :hover and :focus.
* Display in a `::after` pseudo-element, with the text
  inserted by CSS using `content: attr(aria-description);`.
* Style the element by applying a Codex mixin, so that no
  separate module is needed, and no separate HTML element that
  applies the cdx-tooltip class. Instead, the Codex style can
  be applied as a mixin to our selector directly
  ```
  @import ( less ) 'mediawiki.skin.codex/modules/CdxTooltip.css';
  mw-tempuserlink-expired::after {
    .cdx-tooltip;
    content: attr(aria-description);
    position: absolute;
    top: 100%;
    left: 0;
    display: none;
  }
  mw-tempuserlink-expired:hover::after,
  mw-tempuserlink-expired:focus::after {
    display: block;
  }
  ```

This would work, but this imports a private file. Afaik there
is not a publicly supported way to import this (yet).

== Next next best thing ==

* Use a stub element, similar to the previous approach.
  This exists to carry the cdx-tooltip class name.

== While at it ==

* I set the `title` attribute to null so that it is removed from
  the DOM instead of set to an empty string.
* The element has role=presentation now, since the accessible
  info is natively in aria-described, and does not rely on the
  visual tooltip.

== Resubmission ==

This is a re-submission of I2688dc1299ad, which we reverted because
CheckUser ipReveal.js depended on recursively scraping text content
from `mw-tempuserlink` elements to extract temp user names, and adding
the tooltip element as a child broke that. This was fixed by adding a
`data-mw-target` attribute in I32a8a01b47f20, which CheckUser uses
instead as of I225b38c1a083ec9.

I've also removed the no-longer-applicable regression test for
`textContext` added in Ic8d9ca1f02, as this no longer passes and
is obsoleted by the data attribute being asserted.

Bug: T391907
Bug: T392522
Change-Id: Ib7512831504aa2b51992588c27aaa196871f79df