media: Minor clarification around touchedCallback
The double condition "$value" and "hook" implies an else branch that
is differnet from our intention. Rather than rejecting the cache
(by returning current time) for cases where we have a cache
hit and the hook rejected, we reject the absence of cache OR the hook
returned false.

That is, it implies that if !$value, we also return time() which is
not our intent, since the hook didn't inform that, the hook only
knows what it got, and whether it wants to reject it.

Also, $value is always set because touchedCallback is not be called
for cache misses. I've flipped it so that now the special case / early
return, is for when hook returned false. The regular case (everything
else) is to return null / change nothing.

Change-Id: Id132993d1490cc455b4b6f5676e91837a5c8a4ea