Add Html::expandClassList helper method
Occasionally we have to pass tag attribute data to a template that
only accepts strings, unlike Html::element() and similar methods
that accept strings as well as arrays in several different formats.
Add a helper that handles all of the formats correctly and converts
them to a string like the Html class, and use it in a few places.

Change-Id: I9b327fd1c6903faa182cf4fe6e8713e5d91c7169