Add Html::addClass helper method
Quite often in MediaWiki we receive an array of tag attributes
(intended to be passed to Html::element() or a similar method)
and want to add another CSS class to the 'class' attribute.
Because space-separated attributes can be passed to Html::element()
as strings as well as arrays in several different formats, doing
this is actually a bit tricky. Add a helper that does it the right
way, and use it in a few places.

Change-Id: I221f66fddd39cec3daa63ae29d6bf336a2d1a762