ResourceLoader: Update testIndirectDependenciesUpdate() to set $IP
== Pattern ==

* The test is modifying the return value before asserting,
  by applying Module::expandRelativePaths().
* The test is using __DIR__ and realpath() to compute the expected
  value, instead of having a fixed and consistent output.

This almost always a bad sign and lowers the confidence of the test,
because we have to assume our computation logic inside the test is
correct, which kinda of justifies its own test. And, we have to assume
that the modification of the actual value is not hiding some problem.

We should test the value directly whenever possible. Now that
getFileDependencies() returns a relative path, this should be easy
to do, since the relative path will not include the location of
your personal MediaWiki installation and thus not vary between people
or between local and CI

== Change ==

The testIndirectDependencies() test is simple and we can remove
the expansion from both the actual and expected value and simply
assert the return value directly.

This other test was storing files outside the MediaWiki install path,
which meant that the relative file ends up including the absolute path
inside the relative path, because there is no overlap between
/where/you/have/mediawiki and /tmp/MW_PHPUnit_123/. This could be
worked around, but this scenario is actually unrealistic and not what
we wanted to test. The test should have included setting $IP so that
we can check how it behaves within the temporary context.

The mistake to not set $IP was not realized when we wrote the test,
because getFileDependencies() returned absolute paths then.

Change I6684dd6faa3c20036 improved this by making them return the
stored paths, in their relative form. When we remove the
Module::expandRelativePaths() call from these in the test, we see
that these paths are indeed like "../../../". But..., we cannot
assert those safely, because the relative path between your MediaWiki
installation and your /tmp directory is greatly dependent on your
operating system and dev environment.

For me:
* /Users/krinkle/Development/mediawiki
* /private/tmp/khc4z6k/MW_PHPUnit_ResourceLoader_a6WBDc
* Actual: ../../../../private/tmp/xxx/khc4z6k/MW_PHPUnit_ResourceLoader_a6WBDc

In CI:
* /workspace/src
* /tmp/MW_PHPUnit_ResourceLoader_mLamQh/
* Actual: ../../tmp/MW_PHPUnit_ResourceLoader_mLamQh

The point of storing relative paths is that they are universal and
deterministic between installations. The bug here was that the
test didn't set $IP to an appropriate base directory, but because we
1) returned absolute paths and 2) used realpath() in CI, it did not
make it harder to test.

Change-Id: I77bed0236a0fd65e3b45110118e62c583f06a4a6