During upload ping db before entering atomic section
Uploads of large files (PublishStashedFile jobs) are failing
because the DB connection is timing out. Normally the connection
would be automatically reconnected, but reconnection does not happen
inside an atomic section. Ping the DB to force a reconnection
prior to doing the atomic stuff inside an upload.

This is meant as a temporary solution until the upload jobs can
be refactored to be smarter about DB transactions.

Bug: T391473
Change-Id: Ia24287aa307a26df8c21b8bcdd96cd424ce6ef2b