block: In block log snippets, link to all logs affecting the user
When showing block log snippets on Special:Contributions, EditPage and
Article::showMissing(), multiple block targets may apply to the given
user, for example if the user is an IP address and there is both a
single IP and range block. A randomly selected target was used for log
snippets and links.

So:

* Pass an array of pages to LogEventsList::showLogExtract()
* In showLogExtract(), in the "show full log" link, pass the array of
  pages as an array parameter to Special:Log.
* Accept an array of pages in LogPager.
* In Special:Log, accept an array of pages in the URL. There is no form
  UI for this and the target box is just blank, but at least the
  correct log events are shown.

I tested the query in production, performance should be fine.

There is no existing test coverage of LogEventsList or LogPager. I
added an integration test for the URL parameters in Special:Log.

Also:

* Deprecate unused public methods of LogPager which make maintenance in
  this area more difficult.
* Deprecated unused hook LogEventsListShowLogExtract.

Bug: T389355
Change-Id: Ie3e8e57f117eb9c2046c32d69bfd5fa441aaf5eb