rdbms: Remove SQLPlatform "stable to override" annotations
You can override a method that is not stable to override. The stability
annotation just introduces a requirement for backwards compatibility and
a deprecation period over a release cycle. "Stable to override" is the
most stringent of stability annotations since it requires that a method
must continue to be called by core until the deprecation process is
complete.

Methods that are not intended to be overridden by extensions should not
be marked "stable to override". We've previously established that there
is no stable interface by which an extension can implement its own
database type.

Change-Id: Iec57af98c37b126e6222b65662dddfc49e4b7585