mime: Use LIBXML_PARSEHUGE during uploaded file validation
It maps to libxml2 XML_PARSE_HUGE option, which relaxes some
arbitrary limits in the XML parser.

In particular, this allows SVG images that contain embedded raster
images over 10 MB to be uploaded.

Apart from increasing potential memory use, the option seems safe and
harmless, and we use it elsewhere in MediaWiki and related code.

Bug: T387969
Change-Id: I019b5c58b0428910ff73e956873e5f3926c28883