block: simplify range queries in DatabaseBlockStore
Multiple places call `getRangeCond()` with multiple ranges, and then OR
the SQL conditions later. This works, but the resulting query is
unnecessarily hard to read, and some parts are duplicated.

So, introduce a new `getConditionForRanges()` method that builds a list
of predicates given a list of ranges. In doing so, it deduplicates some
conditions to produce a more readable query.

Update core callers to use this new method when appropriate. For
`newListFromIPs` in particular, remove duplicates as the very first
thing, so that we don't need an additional exit point later. Also
rename a variable to reflect usage, as it gets overwritten immediately.

Change-Id: Id0826038f2b7a58aa8db51bb80197e96b64257c9