Deprecate calling public methods of SiteConfiguration with incorrect typed args
In T391871 it turns out that Extension:SecurePoll was calling methods
of SiteConfiguration with $wiki=null, which violates the phan-declared types
of the method arguments.

Since these are public APIs, we can't immediately change to declared
and enforced types here.  First hard deprecate calling these methods
with types inconsistent with the declarations; in the next release we
can remove the deprecation warnings and add PHP-enforced type hints
instead.

Bug: T391871
Depends-On: I98d22b018201ad2bafcddb033226586b67bfee8f
Depends-On: If4a42bab31f11113a56b4d707239fc8d0fa9bf46
Change-Id: I8df48a00e1b5c043bed95efdeab42ed99a5b98da