Add user-related link colors to LinkRenderer::getLinkClasses
Refactor UserLinkRenderer: pull it inside of LinkRenderer (to manage
the circular dependency) and deprecate existing direct use: going
forward folks should use LinkRenderer::makeUserLink() instead of
directly accessing the UserLinkRenderer service. Keep the user
link related code in UserLinkRenderer for easier discoverability
since LinkRenderer is already over-stuffed.

LinkRenderer::getLinkClasses() now delegates to UserLinkRenderer
to get the appropriate link colors for link targets pointing to
user-related pages (direct links to user or user talk pages, or
links to Special:Contributions).

At the moment I've just added the link colors which depend only
on the user name string, not the UserIdentity; that is,
'mw-userlink' and 'mw-tempuserlink'. We could refactor to add
the other link colors, but as the number of links on a page can
be very large, it would require batching the UserIdentity lookup
for all of the potential usernames rather that resolving them
individually.  In order to do *that*, we'd need a bulk interface
method (like the GetLinkColours hook) not the individual
one-by-one API currently provided by
LinkRenderer::getLinkClasses(). Reworking the link color API is
put off for a future refactor if there is need.

Bug: T392775
Co-authored-by: Marcin Szwarc <mszwarc@wikimedia.org>
Change-Id: Idfa1e6797581cca8397b65e9539250b1ae60516f