diff.js: don't clear selections when making a new one
removeAllRanges() may end up clearing a selection as a new one is being
made. On undo action=edit pages, this can conflict with integrations
that act on the textarea such as Extension:CodeMirror.

There are side effects that may come with not clearing the selection,
specifically with Shift-clicking to make a selection within one side of
the diff. However this less is a sacrifice worth taking if it conflicts
with editing. A purely CSS solution to copy-and-paste-diffs is on the
horizon once browser support for CSS grid is more widespread (T270775).

Bug: T391168
Change-Id: I19e3dffa6153bd8306627f1a83ccb523cec0e8a6