Upstream editsection wrapping rules from VE
Since time immemorial (or at least since or before this
commit in 2013: Iab9c81222) VE has applied these styles
to editsection links in the init module, i.e. whenever
VE *might* load on the page. In practice this means almost
all article page views on WMF wikis.

This code was probably originally put in VE as VE was
responsible for changing the edit section link from
'edit' to 'edit source' or 'edit | edit source', so prior
to this the link had no spaces in all or most languages.

There is no reason for this rule not to be in core, and
as this is already the case on most pages, this shouldn't
have much effect.

Also applies these styles to .mw-editsectionlike

Bug: T392817
Change-Id: I0e994b1f2be3b467d3564edbcb1c7af9af8f569a