Fix RevisionDelete / EditTags checkboxes
The 'name' attribute for the checkboxes was specified twice (one as
dedicated parameter, once as an entry in the attributes parameter).
In the Xml class (before b9b84e6651afd813ef26064c6a821af25a34d38a),
the latter took priority, but in the Html class, the former does.

Remove the incorrect one. This matches similar code in HistoryPager
and ImageHistoryList.

Bug: T392832
Change-Id: I69a0056c3625112bc9fd8e203e14fb2bb9938b8d