mediawiki.ui: Fix vertical alignment of buttons when in wikitext
This is finally putting a solution onto this alignment issue.
`1.25rem` is equal to `20px`, and that's applying to all our deployed
skins, resulting in minimal accessibility need for a pointer first
button while still providing text zoom only upscaling if needed.
While Flexbox were the more elegant solution, as mentioned on task
we could end up with more side-effects in buttons that feature
nested elements.

Bug: T277365
Change-Id: I92e990204ebcdaa93dcb90d1f753070237d2e6a2