tests: Consistently fake time in HtmlOutputRendererHelperTest
Under increased load, certain tests could fail if they take longer than
1s, because that is the cache TTL set on the helper.

So, introduce a wrapper that fakes time for both a given cache instance,
and also everything else that relies on ConvertibleTimestamp.
Consistently use this helper throughout the test class, so that the two
clocks are guaranteed to remain in sync.

Bug: T342571
Change-Id: I3382ea12f23893c2204caa143e31837e6a3c4433