block: Suppress transaction profiler warning on User::spreadBlock
Move and edit GET requests were scheduling a deferred update which
called User::spreadAnyEditBlock(). If the user was blocked, this would
trigger a transaction profiler warning, since a write query is done on a
GET request.

The feature is important and the latency is acceptable in the post-send
context, so factor out these calls to a new helper which disables the
transaction profiler.

Bug: T391980
Change-Id: I768eeb834dce0f24714f2869514256a75248ca30